moved jpeg header check method to public image utils class.
